Add Debian "stretch" packaging changes for LinOTP admin CLI.
This is a mostly-unremarkable set of changes that adapts the Debian
packaging of the "linotp-adminclient-cli" package to the conventions
of Debian stretch. This lifts the debhelper level from 7 to 9 and the
Debian policy version from 3.9.1 to 3.9.8, and replaces pysupport by
dh-python (pybuild). It also drops support for Python versions older
than 2.7. Tests during package building are disabled for now. In the
package Makefile, the DPKG_BUILDPACKAGE_ARGS variable is passed
through to the dpkg-buildpackage command.

See LINOTP-497, LINOTP-498, LINOTP-499, LINOTP-500, LINOTP-507.

Change-Id: If346b2e9f794f70a49c0de24182bd6a007a93304